+++ src/extension.ts
@@ -1,59 +1,30 @@
 import * as vscode from "vscode";
 import { ProjectNodeProvider, ProjectItem } from "./data";
-import { Config, JSONLang, JSONAppType } from "./config";
+import { Config } from "./config";
+import * as JSON from "./json";
 import * as file from "./file";
 import * as path from "path";
+import { Window } from "./window";
 
-function openFolder(path: string) {
-	let uri = vscode.Uri.file(path);
-	vscode.commands.executeCommand("vscode.openFolder", uri).then(() => {
-		vscode.window.showInformationMessage("Folder " + path + " has been opened. ");
-	});
-}
+function runTask(command: string, folder: string) {
 
-function showInput(input: vscode.InputBoxOptions, check: (value: string | undefined) => boolean): Thenable<string | undefined> {
-	return new Promise<string>((resolve, reject) => {
-		vscode.window.showInputBox(input).then(val => {
-			if (val === undefined) {
-				reject();
-			}
+	let name = path.basename(folder);
+	command = command.split("{workspaceName}").join(name);
+	command = command.split("{workspaceFolder}").join(folder);
 
-			if (check(val)) {
-				resolve(val);
-			} else {
-				showInput(input, check).then(e => resolve(e));
-			}
-		});
-	});
-}
-
-function pickLanguage(config: Config): Thenable<JSONLang> {
-	let langs = config.getProjects().langs;
-	return new Promise<JSONLang>((resolve, reject) => {
-		vscode.window.showQuickPick(langs.map(lang => lang.name)).then(langName => {
-			if (langName === undefined) {
-				reject();
-			}
-
-			let lang = langs.find(lang => lang.name === langName);
-			resolve(lang);
-		});
-	});
-}
-
-function pickProject(config: Config): Thenable<string> {
-	return new Promise<string>((resolve, reject) => {
-		pickLanguage(config).then(lang => {
-			let folders = ["."].concat((file.getFolders(lang.path)));
-			vscode.window.showQuickPick(folders).then(folder => {
-				if (folder === undefined) {
-					reject();
-				}
-
-				resolve(lang.path + "/" + folder);
-			});
-		});
-	});
+	try {
+		let type = "shell";
+		let task = new vscode.Task(
+			{ type: type },
+			vscode.workspace.getWorkspaceFolder(vscode.Uri.file(folder))!,
+			"Project manager", //terminal name
+			type, //source
+			new vscode.ShellExecution(command, { cwd: folder }),
+		);
+		vscode.tasks.executeTask(task);
+	} catch (err) {
+		vscode.window.showInformationMessage(err);
+	}
 }
 
 export function activate(context: vscode.ExtensionContext) {
@@ -64,20 +35,28 @@ export function activate(context: vscode.ExtensionContext) {
 		file.createFolder(root);
 	}
 
-	let config = new Config(root + "/settings.json");
+	let config = new Config(root, context.extensionPath);
 	if (!file.pathExists(root + "/templates")) {
 		file.createFolder(root + "/templates");
 	}
 
+	let window = new Window(config);
 	let provider = new ProjectNodeProvider(config);
 	config.onChangeSettings = () => provider.refresh();
-
+ 
 	vscode.window.registerTreeDataProvider("nodeProjects", provider);
-	vscode.commands.registerCommand("projectmanager.openProject", (arg: string | undefined) => { 
-		if (arg) {
-			openFolder(arg as string);
+	vscode.commands.registerCommand("projectmanager.openProject", (arg1 : string, arg2: string) => { 
+		if (arg1 && arg2) {
+			config.editProject({
+				lang: arg2,
+				path: arg1
+			});
+			window.openFolder(arg1);
 		} else {
-			pickProject(config).then(path => openFolder(path));
+			window.pickProject().then(project => {
+				config.editProject(project);
+				window.openFolder(project.path);
+			});
 		}
 	});
 	vscode.commands.registerCommand("projectmanager.renameProject", (arg: ProjectItem | undefined) => {
@@ -94,17 +73,17 @@ export function activate(context: vscode.ExtensionContext) {
 			}
 
 			function check(data: string | undefined): boolean {
-				if (file.pathExists(folder + data!!)) {
-					vscode.window.showInformationMessage("Project with name '" + data!! + "' already exist. ");
+				if (file.pathExists(folder + data!)) {
+					vscode.window.showInformationMessage("Project with name '" + data! + "' already exist. ");
 					return false;
 				}
 				return true;
 			}
 
-			showInput({ prompt: "Enter new project name" }, check).then(name => {
-				file.renameFolder(source, name!!);
+			window.showInput({ prompt: "Enter new project name" }, check).then(name => {
+				file.renameFolder(source, name!);
 				if (open) {
-					openFolder(folder + name!!);
+					window.openFolder(folder + name!);
 				}
 				provider.refresh();
 			});
@@ -113,7 +92,7 @@ export function activate(context: vscode.ExtensionContext) {
 		if (arg) {
 			rename(arg.tag as string);
 		} else {
-			pickProject(config).then(project => rename(project));
+			window.pickProject().then(project => rename(project.path));
 		}
 	});
 	vscode.commands.registerCommand("projectmanager.deleteProject", (arg: ProjectItem | undefined) => {		
@@ -124,6 +103,7 @@ export function activate(context: vscode.ExtensionContext) {
 			}
 
 			file.deleteFolder(source);
+			config.removeProject(source);
 			provider.refresh();
 			vscode.window.showInformationMessage("Project has been deleted. ");
 		}
@@ -131,12 +111,12 @@ export function activate(context: vscode.ExtensionContext) {
 		if (arg) {
 			deleteProject(arg.tag as string);
 		} else {
-			pickProject(config).then(path => deleteProject(path));
+			window.pickProject().then(project => deleteProject(project.path));
 		}
 	});
 	vscode.commands.registerCommand("projectmanager.config", () => file.openFile(config.path));
 	vscode.commands.registerCommand("projectmanager.removeLanguage", (arg: ProjectItem | undefined) => {
-		function remove(lang: JSONLang) {
+		function remove(lang: JSON.ILang) {
 			config.removeLang(lang);
 			config.save();
 
@@ -145,9 +125,9 @@ export function activate(context: vscode.ExtensionContext) {
 		}
 		
 		if (arg) {
-			remove(arg.tag as JSONLang);
+			remove(arg.tag as JSON.ILang);
 		} else {
-			pickLanguage(config).then(lang => remove(lang));
+			window.pickLanguage().then(lang => remove(lang));
 		}
 	});
 	vscode.commands.registerCommand("projectmanager.addLanguage", () => {
@@ -159,8 +139,8 @@ export function activate(context: vscode.ExtensionContext) {
 			}
 			return true;
 		}
-		showInput({ prompt: "Enter language name" }, check).then(name => {
-			let id = name!!.toLowerCase().replace(" ", "_");
+		window.showInput({ prompt: "Enter language name" }, check).then(name => {
+			let id = name!.toLowerCase().split(" ").join("_");
 			let openFolder = {
 				canSelectMany: false,
 				canSelectFiles: false, 
@@ -170,38 +150,40 @@ export function activate(context: vscode.ExtensionContext) {
 
 			vscode.window.showOpenDialog(openFolder).then(path => {
 				config.addLang({
-					id: id!!, 
-					name: name!!,
-					path: path!![0].fsPath,
+					id: id!, 
+					name: name!,
+					path: path![0].fsPath,
 					types: undefined
 				});
 				
 				config.save();
-				file.createFolder(root + "/templates/" + id!!);
+				file.createFolder(root + "/templates/" + id!);
 			});
 		});
 	});
 	vscode.commands.registerCommand("projectmanager.deleteTemplate", () => {
-		pickLanguage(config).then(lang => {
-			vscode.window.showQuickPick(lang.types!!.map(type => type.name)).then(typeName => {
-				let type = lang.types!!.find(type => type.name === typeName)!!;
-				
-				config.deleteAppType(lang, type);
-				config.save();
+		window.pickLanguage().then(lang => {
+			vscode.window.showQuickPick(lang.types!.map(type => type.name), { placeHolder: "Pick template to delete" }).then(typeName => {
+				if (typeName) {
+					let type = lang.types!.find(type => type.name === typeName)!;
+					
+					config.deleteTemplate(lang, type);
+					config.save();
 
-				file.deleteFolder(root + "/templates/" + lang.id + "/" + type.id);
+					file.deleteFolder(root + "/templates/" + lang.id + "/" + type.id);
+				}
 			});
 		});
 	});
 	vscode.commands.registerCommand("projectmanager.editTemplate", () => {
-		pickLanguage(config).then(lang => {
+		window.pickLanguage().then(lang => {
 			if (lang.types) {
 				if (lang.types.length === 1) {
-					openFolder(root + "/templates/" + lang.id + "/" + lang.types[0].id);
+					window.openFolder(root + "/templates/" + lang.id + "/" + lang.types[0].id);
 				} else {
-					vscode.window.showQuickPick(lang.types.map(type => type.name)).then(typeName =>{
-						let type = lang.types!!.find(type => type.name === typeName)!!;
-						openFolder(root + "/templates/" + lang.id + "/" + type.id);
+					vscode.window.showQuickPick(lang.types.map(type => type.name), { placeHolder: "Pick template to edit" }).then(typeName =>{
+						let type = lang.types!.find(type => type.name === typeName)!;
+						window.openFolder(root + "/templates/" + lang.id + "/" + type.id);
 					});
 				}
 			} else {
@@ -210,9 +192,9 @@ export function activate(context: vscode.ExtensionContext) {
 		});
 	});
 	vscode.commands.registerCommand("projectmanager.createTemplate", () => {
-		pickLanguage(config).then(lang => {
+		window.pickLanguage().then(lang => {
 			function check(data: string | undefined): boolean {
-				let typeId = data!!.toLowerCase().replace(" ", "_");
+				let typeId = data!.toLowerCase().split(" ").join("_");
 				let dirName = typeId;
 				if (data === "") {
 					dirName = "default";
@@ -225,8 +207,8 @@ export function activate(context: vscode.ExtensionContext) {
 				return true;
 			}
 
-			showInput({ prompt: "Enter template name (let empty to default)"}, check).then(template => {
-				let typeId = template!!.toLowerCase().replace(" ", "_");
+			window.showInput({ prompt: "Enter template name (let empty to default)"}, check).then(template => {
+				let typeId = template!.toLowerCase().split(" ").join("_");
 				let name = template;
 
 				if (template === "") {
@@ -234,21 +216,22 @@ export function activate(context: vscode.ExtensionContext) {
 					name = "(default)";
 				}
 
-				let type: JSONAppType = {
+				let type: JSON.ITemplate = {
 					id: typeId,
-					name: name!!
+					name: name!,
+					copyFolder: true
 				};
 
-				config.addAppType(lang.id, type);
+				config.addTemplate(lang.id, type);
 				config.save();
 
 				file.createFolder(root + "/templates/" + lang.id + "/" + typeId);
-				openFolder(root + "/templates/" + lang.id + "/" + typeId);
+				window.openFolder(root + "/templates/" + lang.id + "/" + typeId);
 			});
 		});
 	});
 	vscode.commands.registerCommand("projectmanager.createProject", (arg: ProjectItem | undefined) => {
-		function createProject(lang: JSONLang) {
+		function createProject(lang: JSON.ILang) {
 			function check(data: string | undefined): boolean {
 				if (file.pathExists(lang.path + "/" + data)) {
 					vscode.window.showInformationMessage("Project with name '" + data + "' already exist. ");
@@ -257,27 +240,47 @@ export function activate(context: vscode.ExtensionContext) {
 				return true;
 			}
 
-			function create(type: string = "") {
-				showInput({ prompt: "Enter project name" }, check).then(name => {
-					if (file.pathExists(root + "/templates/" + lang.id + "/" + type)) {
-						file.copyFolder(root + "/templates/" + lang.id + "/" + type, lang.path + "/" + name);
+			function create(type: JSON.ITemplate | undefined = undefined) {
+				window.showInput({ prompt: "Enter project name" }, check).then(name => {
+					if (type) {
+						if ((type.copyFolder === undefined || type.copyFolder) && file.pathExists(root + "/templates/" + lang.id + "/" + type.id)) {
+							file.copyFolder(root + "/templates/" + lang.id + "/" + type.id, lang.path + "/" + name);
+						} else {
+							file.createFolder(lang.path + "/" + name);
+						}
+
+						let source = lang.path + "/" + name;
+						if (type.commands) {
+							type.commands.forEach(command => {
+								runTask(command.script, command.folder ? command.folder : source);
+							});
+						}
 					} else {
 						file.createFolder(lang.path + "/" + name);
 					}
 
-					openFolder(lang.path + "/" + name);
+					config.editProject({
+						lang: lang.id,
+						path: lang.path + "/" + name
+					});
+
 					provider.refresh();
+					vscode.window.showInformationMessage("Open project?", "YES", "NO").then(val => {
+						if (val === "YES") {
+							window.openFolder(lang.path + "/" + name);
+						}
+					});
 				});
 			}
 
 			if (lang.types) {
 				if (lang.types.length > 1) {
 					vscode.window.showQuickPick(lang.types.map(type => type.name)).then(typeName => {
-						let type = lang.types!!.filter(type => type.name === typeName)[0];
-						create(type.id);
+						let type = lang.types!.find(type => type.name === typeName);
+						create(type);
 					});
 				} else {
-					create(lang.types[0].id);
+					create(lang.types[0]);
 				}
 			} else {
 				create();
@@ -285,9 +288,9 @@ export function activate(context: vscode.ExtensionContext) {
 		}
 		
 		if (arg) {
-			createProject(arg.tag as JSONLang);
+			createProject(arg.tag as JSON.ILang);
 		} else {
-			pickLanguage(config).then(lang => createProject(lang));
+			window.pickLanguage().then(lang => createProject(lang));
 		}
 	});
 }
